requirements.txt
pyyaml
python-dotenv
requests
mysql-connector-python
pandas
pytest
pytest-cov
pre-commit
ruff